feat(kvrocks2redis): add Redis Cluster write target support
Introduce ClusterDirectWriter, a Writer subclass that routes RESP
commands directly to cluster nodes without an intermediate AOF file.
Parse each command's key via ExtractRespKeyForRouting (zero-copy CRC16
slot lookup), buffer per node, and flush in pipeline batches of 128.
Write() is synchronous so next_seq_ only advances after the target
cluster confirms every command in the batch, giving crash-safe replay
for free without a separate offset checkpoint file.

On transport failure the topology is refreshed so the next restart
reconnects with an up-to-date node map; the unadvanced next_seq_ then
replays the same WAL batch with the fresh topology.

Also:
- Add ExtractRespKeyForRouting + RespParseResult to string_util
- Fix sync.cc stoi→stoull sequence-number overflow
- Add unit tests for the RESP key extractor
- main.cc branches on target_cluster_enabled to select the writer

+++ utils/kvrocks2redis/sync.cc
@@ -183,7 +183,7 @@ Status Sync::readNextSeqFromFile(rocksdb::SequenceNumber *seq) {
   char buf[22];
   memset(buf, '\0', sizeof(buf));
   if (read(next_seq_fd_, buf, sizeof(buf)) > 0) {
-    *seq = static_cast<rocksdb::SequenceNumber>(std::stoi(buf));
+    *seq = static_cast<rocksdb::SequenceNumber>(std::stoull(buf));
   }
 
   return Status::OK();
